+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/snx-link/snx-link-smart-contracts/compare/v1.0.0...HEAD)
 
+### Added
+
+- Add Guaranteed Autoclaim burning sUSD
+- Add upgradability platform and claimer fees (only decrease)
+- Add upgradability gasOffsetCorrection (only decrease)
+
 ## [1.0.0](https://github.com/snx-link/snx-link-smart-contracts/compare/224a03c6915ccc9f08ee1c03c289915b65bbcf5c...v1.0.0)
 
 - Initial SNXLink v1 release.